requirements.txt
pytorch
nltk
numpy
pyTelegramBotAPI
psycopg2